优化了一些功能

+++ pkg/redis/cluster.go
@@ -3,27 +3,55 @@ package redis
 import (
 	"errors"
 	"fmt"
+	"strconv"
 	"strings"
 
 	"github.com/go-redis/redis/v8"
 )
 
-//GetClusterNodes 获取集群 info 信息
+// GetClusterNodes 获取集群 info 信息
 func GetClusterNodes(addr string, password string) (ret string, err error) {
 	// init showSlowLog cluster conn
 	var rc *redis.ClusterClient
-	rc, err = InitClusterRedis(addr, password)
+	rc, err = InitClusterRedis([]string{addr}, password)
 	defer rc.Close()
 
 	// showSlowLog command: cluster nodes
-	ret, err = rc.ClusterNodes(CTX).Result()
+	ret, err = rc.ClusterNodes(ctx).Result()
 	if err != nil {
 		return "", err
 	}
 	return
 }
 
-//FormatClusterNodes 获取集群节点信息,并按 master/slave 进行分类:{"master":[], "slave":[]}
+// CheckClusterConfig 校验集群配置项是否一致
+func CheckClusterConfig(addrSlice []string, password, configArg string) (bool, error) {
+	var retValue string
+	for _, addr := range addrSlice {
+		// 创建 redis 连接
+		rc, err := InitStandRedis(addr, password)
+		if err != nil {
+			return false, err
+		}
+
+		// 获取配置项
+		argRet, err := rc.ConfigGet(ctx, configArg).Result()
+		if err != nil {
+			return false, err
+		}
+
+		if retValue != argRet[1].(string) && retValue != "" {
+			err = errors.New("集群配置项的值不一致")
+			return false, err
+		} else {
+			retValue = argRet[1].(string)
+		}
+		rc.Close()
+	}
+	return true, nil
+}
+
+// FormatClusterNodes 获取集群节点信息,并按 master/slave 进行分类:{"master":[], "slave":[]}
 func FormatClusterNodes(addr string, password string) (data *ClusterNodesMap, err error) {
 	var ClusterNodesDetail []*ClusterNode
 	var MasterInstances []string
@@ -50,6 +78,7 @@ func FormatClusterNodes(addr string, password string) (data *ClusterNodesMap, er
 		Addr := strings.Split(nodeSlice[1], "@")[0]
 		IDToAddr[ID] = Addr
 		AddrToID[Addr] = ID
+		SlotStr := strings.Split(node, "connected")[1]
 
 		role := nodeSlice[2]
 		if strings.Contains(role, "myself") {
@@ -61,10 +90,12 @@ func FormatClusterNodes(addr string, password string) (data *ClusterNodesMap, er
 			if _, ok := NodeTmpMap[nodeSlice[0]]; !ok {
 				NodeTmpMap[nodeSlice[0]] = map[string]string{
 					"masterAddr": Addr,
+					"SlotStr":    strings.Trim(SlotStr, " "),
 				}
 				continue
 			}
 			NodeTmpMap[nodeSlice[0]]["masterAddr"] = Addr
+			NodeTmpMap[nodeSlice[0]]["SlotStr"] = strings.Trim(SlotStr, " ")
 		} else if role == "slave" {
 			SlaveInstances = append(SlaveInstances, Addr)
 			if _, ok := NodeTmpMap[nodeSlice[3]]; !ok {
@@ -89,6 +120,7 @@ func FormatClusterNodes(addr string, password string) (data *ClusterNodesMap, er
 			item["masterAddr"],
 			item["slaveAddr"],
 			item["slaveID"],
+			item["SlotStr"],
 		}
 		ClusterNodesDetail = append(ClusterNodesDetail, node)
 	}
@@ -102,7 +134,7 @@ func FormatClusterNodes(addr string, password string) (data *ClusterNodesMap, er
 	return
 }
 
-//ClusterGetConfig 获取集群配置并校验是否一致
+// ClusterGetConfig 获取集群配置并校验是否一致
 func ClusterGetConfig(addrSlice []string, password string, configArg string) (ret string, err error) {
 	for _, addr := range addrSlice {
 		// init showSlowLog conn
@@ -111,7 +143,7 @@ func ClusterGetConfig(addrSlice []string, password string, configArg string) (re
 			return "", err
 		}
 
-		argRet, err := rc.ConfigGet(CTX, configArg).Result()
+		argRet, err := rc.ConfigGet(ctx, configArg).Result()
 		if err != nil {
 			return "", err
 		}
@@ -127,7 +159,7 @@ func ClusterGetConfig(addrSlice []string, password string, configArg string) (re
 	return
 }
 
-//ClusterSetConfig 批量设置集群配置
+// ClusterSetConfig 批量设置集群配置
 func ClusterSetConfig(addrSlice []string, password string, configArg string, setValue string) (err error) {
 	for _, addr := range addrSlice {
 		// init showSlowLog conn
@@ -136,7 +168,7 @@ func ClusterSetConfig(addrSlice []string, password string, configArg string, set
 			return err
 		}
 
-		err = rc.ConfigSet(CTX, configArg, setValue).Err()
+		err = rc.ConfigSet(ctx, configArg, setValue).Err()
 		if err != nil {
 			return err
 		}
@@ -145,9 +177,17 @@ func ClusterSetConfig(addrSlice []string, password string, configArg string, set
 	return
 }
 
-//ClusterFlushAll 集群数据清理
-func ClusterFlushAll(ClusterNodesSlice []string, password string, flushCMD string) (err error) {
-	for _, addr := range ClusterNodesSlice {
+// ClusterFlushAll 集群数据清理
+func ClusterFlushAll(data *ClusterNodesMap, password string, flushCMD string) (err error) {
+	clusterNodes := append(data.Masters, data.Slaves...)
+	// 获取cluster-node-timeout配置值
+	ret, err := ClusterGetConfig(clusterNodes, password, "cluster-node-timeout")
+	if err != nil {
+		fmt.Printf("获取集群配置项cluster-node-timeout失败, err:%v\n", err)
+		return
+	}
+
+	for _, addr := range clusterNodes {
 		// init showSlowLog conn
 		rc, err := InitStandRedis(addr, password)
 		if err != nil {
@@ -155,16 +195,45 @@ func ClusterFlushAll(ClusterNodesSlice []string, password string, flushCMD strin
 		}
 		defer rc.Close()
 
-		//对每个节点执行 FLUSHALL 命令
-		if flushCMD == "FLUSHALL" {
-			err = rc.FlushAll(CTX).Err()
+		// 获取 redis 版本
+		infoMap, err := FormatRedisInfo(addr, password)
+		versionPrefixStr := strings.Split(infoMap["redis_version"], ".")[0]
+		versionPrefix, err := strconv.ParseInt(versionPrefixStr, 10, 64)
+		if err != nil {
+			fmt.Printf("获取redis: %s 版本失败\n", addr)
+			return
+		}
+
+		// 针对不同版本的 redis, 执行不同的的清空操作
+		if versionPrefix == 3 { // 清空会堵塞 redis,造成主从切换,需要先调整集群超时时间
+			// 调整将cluster-node-timeout配置项的值为 30 分钟,避免清空 redis 的时候发生主从切换
+			err = ClusterSetConfig(clusterNodes, password, "cluster-node-timeout", ret)
 			if err != nil {
+				fmt.Printf("还原集群配置项cluster-node-timeout失败,配置项初始值为%s, err:%v\n", ret, err)
+				return
+			}
+
+			//对每个节点执行 FLUSHALL 命令
+			if flushCMD == "FLUSHALL" {
+				err = rc.FlushAll(ctx).Err()
+				if err != nil {
+					return err
+				}
+			} else {
+				err = rc.Do(ctx, flushCMD).Err()
 				return err
 			}
-		} else {
-			err = rc.Do(CTX, flushCMD).Err()
-			return err
+
+			// 将cluster-node-timeout配置修改为原来配置的值
+			err = ClusterSetConfig(clusterNodes, password, "cluster-node-timeout", ret)
+			if err != nil {
+				fmt.Printf("还原集群配置项cluster-node-timeout失败,配置项初始值为%s, err:%v\n", ret, err)
+				return
+			}
+		} else if versionPrefix >= 4 { // 执行异步清空
+
 		}
+
 	}
 	return
 }
